requirements.txt
tensorflow
flask
numpy
pillow
requests
streamlit
gunicorn